// src/types/vnpay.d.ts
export interface VnpReturnUrlParams {
    vnp_Version: string;
    vnp_Command: string;
    vnp_CurrCode: string;
    vnp_IpAddr: string;
    vnp_Locale: string;
    vnp_OrderType: string;
    vnp_TmnCode: string;
    vnp_Amount: number;
    vnp_CreateDate: string;
    vnp_OrderInfo: string;
    vnp_ReturnUrl: string;
    vnp_ExpireDate: string;
    vnp_TxnRef: string;
}